+++ scripts/conab/fetch_reference_samples.sh
@@ -21,6 +21,8 @@ fetch "PrecosSemanalUF.txt" "https://portaldeinformacoes.conab.gov.br/downloads/
 fetch "PrecosSemanalMunicipio.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/PrecosSemanalMunicipio.txt"
 fetch "PrecosMensalUF.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/PrecosMensalUF.txt"
 fetch "PrecosMensalMunicipio.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/PrecosMensalMunicipio.txt"
+fetch "ProhortDiario.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/ProhortDiario.txt"
+fetch "ProhortMensal.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/ProhortMensal.txt"
 fetch "Frete.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/Frete.txt"
 fetch "LevantamentoCana.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/LevantamentoCana.txt"
 fetch "SerieHistoricaCana.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/SerieHistoricaCana.txt"